library-ai-system/api/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
qdrant-client
openai
langchain
langchain-openai
langchain-community
PyPDF2
tiktoken
pydantic[email]
python-multipart
pyjwt
loguru
aiofiles
python-dateutil
pandas
numpy